feat: optimize apis for cuda feature detection.
