+++ LLama.Examples/Program.cs
@@ -7,6 +7,7 @@ Console.WriteLine(" __       __                                       ____     _
 
 Console.WriteLine("======================================================================================================");
 
+NativeLibraryConfig.Default.WithCuda().WithLogs();
 
 NativeApi.llama_empty_call();
 Console.WriteLine();
